pgmpyro_plugins/requirements.txt
pyrogram==2.0.102
TgCrypto==1.2.5
Pillow>=8.4.0
pytz>=2021.3
PyYAML>=6.0
coloredlogs>=15.0.1
psutil>=5.8.0
httpx~=0.23.3
apscheduler>=3.10.1
sqlitedict~=2.1.0
casbin==1.18.0
sentry-sdk==1.16.0
PyQRCode>=1.2.1
PyPng
fastapi==0.94.1
amis-python==1.0.7
python-jose
uvicorn
pydantic==1.10.6
starlette==0.26.1
googlesearch-python==1.1.0
requests==2.28.2
beautifulsoup4==4.11.2
APScheduler==3.10.1
convopyro==0.5
GitPython==3.1.31
uvloop==0.17.0